requirements.txt
streamlit>=1.25.0
google-generativeai>=0.3.1
odfpy>=1.4.1
Pillow>=9.4.0
pandas>=1.5.0
numpy>=1.24.0